requirements.txt
asgiref==3.5.0
click==8.1.2
colorama==0.4.4
Django==3.2.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djlint==0.7.6
importlib-metadata==4.11.3
pathspec==0.9.0
PyJWT==2.3.0
pytz==2022.1
PyYAML==6.0
regex==2022.3.15
sqlparse==0.4.2
tomlkit==0.9.2
tqdm==4.64.0
zipp==3.8.0
